+++ src/sys/os/linux/consts/mod.rs
@@ -8,9 +8,10 @@ pub(super) mod file; // LINUX_[FD|O_FLAGS].
 pub(super) mod ioctl; // LINUX_IOCTL.
 pub(super) mod signal; // LINUX_[SIGACTION|SIGNAL].
 pub(super) mod termios; // LINUX_TERMIOS_[I|O|C|L]FLAG.
+pub(super) mod time; // LINUX_TIME
 
 crate::structural_mods! { // _mods
     _mods {
-        pub use super::{errno::*, file::*, ioctl::*, signal::*, termios::*};
+        pub use super::{errno::*, file::*, ioctl::*, signal::*, termios::*, time::*};
     }
 }