+++ src/sys/os/linux/syscalls/riscv.rs
@@ -355,6 +355,22 @@ impl Linux {
 /// # Syscalls: Timing and signal handling.
 impl Linux {
     #[must_use]
+    #[doc = SYS_CLOCK_GETTIME!()]
+    pub unsafe fn sys_clock_gettime(clock_id: c_int, tp: *mut LinuxTimespec) -> isize {
+        unsafe {
+            asm!(
+                "li a7, {CLOCK_GETTIME}",
+                "ecall",
+                CLOCK_GETTIME = const SYS::CLOCK_GETTIME,
+                in("a0") clock_id,
+                in("a1") tp,
+                lateout("a0") result,
+                options(nostack)
+            );
+            result
+        }
+    }
+    #[must_use]
     #[doc = SYS_NANOSLEEP!()]
     pub unsafe fn sys_nanosleep(req: *const LinuxTimespec, rem: *mut LinuxTimespec) -> isize {
         let result;